# frontend/public/openapi.yaml
openapi: 3.0.3
info:
  title: StrikeGod 排行榜对外查询 API
  version: "1.0.0"
  description: |
    排行榜查询服务的**对外公开**只读接口。无需鉴权，直接 GET 即可。

    ## 两套「区服编号」务必区分（最常见的坑）

    游戏里一个区服有两个编号，本 API 不同端点用的是**不同**的那一个：

    | 端点 | `server` 用哪种编号 | 取值来源 |
    |------|--------------------|----------|
    | `/servers`、`/tab`、`/board` | **登录 id（loginId）** | 即 `/api/public/servers` 返回的 `id` 字段 |
    | `/player` | **对外服号（districtId）** | 玩家在游戏客户端里看到的服号 |

    > 同一个区，loginId 与 districtId 数值通常**不一样**。例如「联通四区」loginId=10、districtId=16。
    > 如果把 loginId 当 districtId 传给 `/player`，会查到**别的区**或返回「未知区服」。

    ### 区服对照表（loginId ↔ districtId ↔ 区服名）

    | loginId | 区服名 | districtId |
    |--------:|--------|-----------:|
    | 1  | 双线一区 | 5  |
    | 2  | 电信一区 | 6  |
    | 3  | 联通一区 | 7  |
    | 4  | 电信二区 | 8  |
    | 5  | 联通二区 | 9  |
    | 6  | 电信三区 | 10 |
    | 7  | 电信四区 | 12 |
    | 8  | 联通三区 | 13 |
    | 9  | 电信五区 | 14 |
    | 10 | 联通四区 | 16 |
    | 11 | 电信六区 | 17 |
    | 12 | 双线二区 | 19 |
    | 13 | 电信七区 | 22 |
    | 14 | 联通五区 | 18 |
    | 15 | 电信八区 | 24 |
    | 16 | 联通六区 | 28 |
    | 17 | 双线三区 | 31 |
    | 18 | 双线四区 | 42 |
    | 19 | 双线五区 | 45 |
    | 21 | 双线七区 | 52 |

    ## 名次「过滤不变」语义

    `/tab` 的 `q`（模糊搜名）与 `vip`（VIP 等级）过滤，名次都是**先按整榜算好再过滤**——
    所以搜出来的某个人，名次仍是 45 名而不会因过滤变成第 1 名。

    ## 行数据 `rows` 是动态列

    `rows[]` 里每个对象的字段**由同一响应的 `fields[]` 决定**（`fields[].key` 即列名）。
    不同榜列不一样，请按 `fields` 动态渲染，不要写死列名。`fields[].render` 是建议的展示类型。
servers:
  - url: https://example.com
    description: 生产环境
...
paths:
  /api/auth/register:
    post:
      tags: [账号]
      summary: 注册站点账号（注册即登录）
      description: |
        开放注册（Cloudflare Turnstile 拦机器人，启用时需带 `cfToken`）。成功直接发 30 天会话：
        响应体含 `token`，同时写 HttpOnly cookie `user_token`。
        带有效的管理员邀请码（后台签发，`inv_` 开头、一次性、7 天有效）则注册为**管理员**。
        业务失败统一 `{ok:false, message}`（HTTP 200）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, description: 2-24 位，中文/字母/数字/下划线/连字符 }
                password: { type: string, description: 8-72 位 }
                inviteCode: { type: string, description: 管理员邀请码（可选） }
                cfToken: { type: string, description: Cloudflare Turnstile 挑战 token（人机验证二选一；配置见 /api/public/captcha-config） }
                gtPayload: { type: string, description: 极验 GeeTest v4 getValidate() 的 JSON 串（人机验证二选一，与 cfToken 任一即可） }
      responses:
        "200":
          description: 注册结果
          content:
            application/json:
              example: { ok: true, token: "su_xxx", username: "示例", role: "USER" }
  /api/auth/login:
    post:
      tags: [账号]
      summary: 登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200":
          description: 登录结果（成功含 token + 写 cookie；失败 `{ok:false, message}`）
          content:
            application/json:
              example: { ok: true, token: "su_xxx", username: "示例", role: "USER" }
...
  /api/public/weapon/search:
    get:
      tags: [武器]
      summary: 武器 + 配件/技能/词缀 统一检索
      description: |
        对**武器**与**辅助条目**（武器技能 / 配件槽位 / 配件 / 配件技能 / 重构词缀 / 武器属性面板）做
        **全字符串模糊**检索，合并为一个分页列表（武器在前），每页 **100**。
        条目 `type=entry` 含 `payload`（该记录完整原始 JSON 字符串，可解析后展示全部字段）。
      parameters:
        - name: q
          in: query
          required: false
          description: 关键字，对每条记录的**所有字符串字段**（去 HTML）做 ILIKE 模糊匹配。留空返回空结果。
          schema: { type: string }
          example: 姜维
        - name: page
          in: query
          required: false
          description: 页码，从 1 开始；每页 100。
          schema: { type: integer, minimum: 1, default: 1 }
          example: 1
      responses:
        "200":
          description: 合并分页结果
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WeaponSearchPage" }
  /api/public/weapon/{id}:
    get:
      tags: [武器]
      summary: 武器详情（基础 + 面板）
      parameters:
        - name: id
          in: path
          required: true
          description: 武器 id（来自 `/weapon/search` 结果里 `type=weapon` 项的 `id`）。
          schema: { type: integer, format: int64 }
          example: 2038
      responses:
        "200":
          description: 武器详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Weapon" }
        "404":
          description: 未找到该武器
...
components:
  schemas:
...
    WeaponSearchPage:
      type: object
      description: 武器+条目统一搜索分页
      properties:
        total: { type: integer, description: 武器命中数 + 条目命中数 }
        page: { type: integer }
        size: { type: integer, description: 固定 100 }
        items:
          type: array
          items:
            type: object
            description: "type=weapon 为武器精简项；type=entry 为辅助条目（含完整 payload）"
            properties:
              type: { type: string, enum: [weapon, entry] }
              id: { type: integer, format: int64, description: "武器 id（type=weapon）" }
              name: { type: string }
              codeName: { type: string, description: "武器代号（type=weapon）" }
              typeLabel: { type: string, description: "武器类型中文名（type=weapon）" }
              catLabel: { type: string, description: "条目类别名（type=entry），如 武器技能/配件/重构词缀/武器属性" }
              key: { type: string, description: "条目主键（type=entry）" }
              detail: { type: string, description: "条目摘要/面板文本（type=entry）" }
              payload: { type: string, description: "条目完整原始 JSON 字符串（type=entry），解析后可展示全部字段" }
    Weapon:
      type: object
      description: 武器详情（基础属性 + 面板）
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, description: 武器名 }
        codeName: { type: string, description: 代号 }
        type: { type: integer }
        subType: { type: integer }
        typeLabel: { type: string, description: 类型中文名（步枪/狙击/冲锋枪…） }
        description: { type: string, description: 描述（可能含 HTML） }
        specific: { type: string, description: 特性文本 }
        tags: { type: string, description: 标签 }
        icon: { type: string, description: 图标资源 }
        penetrate: { type: number, description: 穿透 }
        fireRate: { type: number, description: 射速 }
        power: { type: number, description: 威力 }
        accurate: { type: number, description: 精准 }
        stability: { type: number, description: 稳定 }
        weight: { type: number, description: 便携 }
        bullet: { type: number, description: 弹匣 }
        bulletMax: { type: number, description: 弹匣上限 }